change dir name from master to masters.
