+++ app/backend/app/Models/Masters/ProductType.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Models\Master;
+namespace App\Models\Masters;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Manufacturers extends Model
+class ProductType extends Model
 {
     use HasFactory;
     use SoftDeletes;
@@ -15,14 +15,12 @@ class Manufacturers extends Model
     public const ID                = 'id';
     public const NAME              = 'name';
     public const DETAIL            = 'detail';
-    public const ADDRESS           = 'address';
-    public const TEL               = 'tel';
     public const CRREATED_AT       = 'created_at';
     public const UPDATED_AT        = 'updated_at';
     public const DELETED_AT        = 'deleted_at';
 
     //テーブル名指定
-    protected $table = 'manufacturers';
+    protected $table = 'product_types';
 
     /**
      * Indicates if the model should be timestamped.
@@ -53,8 +51,6 @@ class Manufacturers extends Model
     protected $fillable = [
         self::NAME,
         self::DETAIL,
-        self::ADDRESS,
-        self::TEL,
         self::CRREATED_AT,
         self::UPDATED_AT
     ];
