+++ app/backend/database/seeders/BaseSeeder.php
@@ -6,7 +6,6 @@ use Illuminate\Database\Seeder;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Config;
-use App\Models\Master\Products;
 
 class BaseSeeder extends Seeder
 {
