+++ app/backend/database/seeders/DatabaseSeeder.php
@@ -17,9 +17,9 @@ class DatabaseSeeder extends Seeder
     {
         // \App\Models\User::factory(10)->create();
         $this->call(AdminsTableSeeder::class);
-         // $this->call(AdminsLogTableSeeder::class);
-        $this->call(Master\ManufacturersTableSeeder::class);
-        $this->call(Master\ProductsTableSeeder::class);
+         // $this->call(Logs\AdminsLogTableSeeder::class);
+        $this->call(Masters\ManufacturersTableSeeder::class);
+        $this->call(Masters\ProductsTableSeeder::class);
         $this->call(PermissionsTableSeeder::class);
         $this->call(RolesTableSeeder::class);
         $this->call(RolePermissionsTableSeeder::class);
