+++ app/backend/database/seeders/masters/ProductsTableSeeder.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Database\Seeders\Master;
+namespace Database\Seeders\Masters;
 
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Config;
-use App\Models\Master\Products;
+use App\Models\Masters\Products;
 use Database\Seeders\BaseSeeder;
 
 class ProductsTableSeeder extends BaseSeeder
